refactor: move internal/autoupdate under internal/app
